maven configuration with sarah 29th may 2024 (#40)
* maven configuration with sarah 29th may 2024

* Slight change to solution

Sophie has run into the same issue that I have had when trying to run mvn compile on the process folder. Her solution is correct but I would change things slightly. Her fix was to update the java version to 22. Which would be fine fine, but I have checked the available java versions on myriad and they don't have this version so we'll have to change the solution slightly.

I have changed the java version back to 1.8, but removed the 'release' flag in the pom.xml file which was causing the issue

---------

Co-authored-by: RobertManningSmith <58108110+RobertManningSmith@users.noreply.github.com>