Big ugly commit...
I've been working on modularizing the infection and infectious behaviour stuff. So far it's working. I've updated a lot of the logging to iterate over the infecitons rather than the person objects so that is good, I didn't finish doing the testing and dummy disease stuff so tests are still failing there...